Fixed bug in SafeAbstractTrader that was causing the wrong orders to be tracked during placement
git-svn-id: http://macdonald.dcs.gla.ac.uk/~tws/repositories/jagora/trunk@52 ef8dee83-5f3d-43c0-baf0-9562671df143

+++ api/src/main/java/uk/ac/glasgow/jagora/Market.java
@@ -25,11 +25,11 @@ public interface Market {
 
 	public List<SellOrder> getSellOrders();
 
-	public Double getBestBidPrice();
+	public Long getBestBidPrice();
 	
-	public Double getBestOfferPrice();
+	public Long getBestOfferPrice();
 	
-	public Double getLastKnownBestBidPrice();
+	public Long getLastKnownBestBidPrice();
 	
-	public Double getLastKnownBestOfferPrice();
+	public Long getLastKnownBestOfferPrice();
 }
