+++ api/src/main/java/uk/ac/glasgow/jagora/Order.java
@@ -17,7 +17,7 @@ public interface Order {
 	public abstract void rollBackTrade(TickEvent<Trade> trade)
 			throws TradeExecutionException;
 
-	public abstract Double getPrice();
+	public abstract Long getPrice();
 	
 	public Boolean isFilled();
