+++ api/src/main/java/uk/ac/glasgow/jagora/trader/Trader.java
@@ -17,7 +17,7 @@ public interface Trader {
 	/**
 	 * @return the cash currently held by this trader.
 	 */
-	public abstract Double getCash();
+	public abstract Long getCash();
 	
 	/**
 	 * @param stock
